+++ packages/prisma-extension-bark/src/create/create-root.js
@@ -6,15 +6,15 @@ import { increment_path, int2str } from '../utils.js'
  * @template A - Args
  *
  * @this {T}
- * @param {import('$types/create').createRootArgs<T, A>} args
- * @returns {Promise<import('$types/create').createRootResult<T, A>>}
+ * @param {import('$types/create.d.ts').createRootArgs<T, A>} args
+ * @returns {Promise<import('$types/create.d.ts').createRootResult<T, A>>}
  */
 export default async function ({ data, ...args }) {
 	const ctx = Prisma.getExtensionContext(this)
 
 	const last_root = await ctx.findLastRoot({ select: {
 		path: true
-	} })
+	}})
 
 	let new_path = null
 	if (last_root) {