+++ packages/prisma-extension-bark/src/delete/delete-node.js
@@ -1,37 +1,32 @@
 import { Prisma } from '@prisma/client'
-import { path_from_depth } from '../utils.js'
+import { has_nullish, path_from_depth } from '../utils.js'
 
 /**
  * @template T - Model
  * @template A - Args
  *
  * @this {T}
- * @param {import('$types/delete').deleteNodeArgs<T, A>} args
- * @returns {Promise<import('$types/delete').deleteNodeResult<T, A>>}
+ * @param {import('$types/delete.d.ts').deleteNodeArgs<T, A>} args
+ * @returns {Promise<import('$types/delete.d.ts').deleteNodeResult<T, A>>}
  */
-export default async function ({ node, where }) {
+export default async function ({ node }) {
 	const ctx = Prisma.getExtensionContext(this)
 
 	/** @type {string} */
-	let path
+	let path = node?.path
 	/** @type {number} */
-	let depth
+	let depth = node?.depth
 
-	if (node) {
-		path = node.path
-		depth = node.depth
-	} else if (where) {
+	if (has_nullish(path, depth)) {
 		const target = await ctx.findUniqueOrThrow({
-			where,
+			where: node,
 			select: {
 				path: true,
 				depth: true
 			}
 		})
-		if (target) {
-			path = target.path
-			depth = target.depth
-		}
+		path = target.path
+		depth = target.depth
 	}