+++ packages/prisma-extension-bark/src/find/find-descendants.js
@@ -1,39 +1,41 @@
 import { Prisma } from '@prisma/client'
 import { default_order_by } from '../consts.js'
+import { has_nullish, merge_where_args } from '../utils.js'
 
 /**
  * @template T - Model
  * @template A - Args
  *
  * @this {T}
- * @param {import('$types/find').findDescendantsArgs<T, A>} args
- * @returns {Promise<import('$types/find').findDescendantsResult<T, A>>}
+ * @param {import('$types/find.d.ts').findDescendantsArgs<T, A>} args
+ * @returns {Promise<import('$types/find.d.ts').findDescendantsResult<T, A>>}
  */
 export default async function ({ node, where, orderBy = default_order_by, ...args }) {
 	const ctx = Prisma.getExtensionContext(this)
 
 	/** @type {string} */
-	let path
+	let path = node?.path
 	/** @type {number} */
-	let depth
+	let depth = node?.depth
 	/** @type {number} */
-	let numchild
-	/** @type {number} */
-	let id
+	let numchild = node?.numchild
 
-	// Get required arguments from instance
-	if (node) {
-		path = node.path
-		depth = node.depth
-		numchild = node.numchild
-		id = node.id
-	} else if (where) {
-		const target = await ctx.findUniqueOrThrow({ where })
+	// Check if all requirements are available
+	if (has_nullish(path, depth, numchild)) {
+		const target = await ctx.findUnique({
+			where: node,
+			select: {
+				path: true,
+				depth: true,
+				numchild: true
+			}
+		})
 		if (target) {
 			path = target.path
 			depth = target.depth
 			numchild = target.numchild
-			id = target.id
+		} else {
+			return null
 		}
 	}
 
@@ -44,17 +46,15 @@ export default async function ({ node, where, orderBy = default_order_by, ...arg
 	}
 
 	return ctx.findMany({
-		where: {
-			id: {
-				not: id
-			},
+		where: merge_where_args({
 			path: {
 				startsWith: path,
+				not: path
 			},
 			depth: {
 				gte: depth
 			}
-		},
+		}, where),
 		orderBy,
 		...args
 	})