+++ packages/prisma-extension-bark/src/find/find-last-root.js
@@ -1,21 +1,23 @@
 import { Prisma } from '@prisma/client'
+import { merge_where_args } from '../utils.js'
+
 
 /**
  * @template T - Model
  * @template A - Args
  *
  * @this {T}
- * @param {import('$types/find').findLastRootNodeArgs<T, A>} args
- * @returns {Promise<import('$types/find').findLastRootNodeResult<T, A>>}
+ * @param {import('$types/find.d.ts').findLastRootNodeArgs<T, A>} args
+ * @returns {Promise<import('$types/find.d.ts').findLastRootNodeResult<T, A>>}
  */
 export default async function (args) {
 	const ctx = Prisma.getExtensionContext(this)
 
 	return ctx.findFirst({
-		...args,
-		where: {
+		where: merge_where_args({
 			depth: 1
-		},
+		}, args?.where),
+		...args,
 		orderBy: {
 			path: 'desc'
 		}